chore: split of submodules from `instruction.rs`
